### Description:
_<Describe proposed changes here...>_

### PR Checklist:
- [ ] PR title prefix added. (ICM, S360, Hot-Fix, Cherry-pick etc.)
- [ ] Dependent changes addressed. (Geneva config, Service Basics, Partner check-ins etc.)
- [ ] Integration tests added/updated.

### Testing:
- [ ] Manual E2E Testing. (Local, DF, Dev, CI, Perf, X-Region, X-Tenant)
- [ ] Automated E2E Testing. (Dev, UX)

## Type of change
- [ ] Bug fix _(non-breaking change which fixes an issue)_
- [ ] New feature _(non-breaking change which adds functionality)_
- [ ] Breaking change _(fix or feature that would cause existing functionality to be altered or to not work as expected)_
- [ ] Infrastructure/Configuration

### Additional Details (optional)
_<Add more details as necessary to help reviewers better understand the changes...>_

Delete root folder bug fix

Related work items: #3453977

+++ src/DataDomain/DimensionalModel/src/main/scala/com/microsoft/azurepurview/dataestatehealth/dimensionalmodel/main/DimensionalModelMain.scala
@@ -24,8 +24,12 @@ object DimensionalModelMain {
     if (fileExists(filePath)) {
       println("SKIP: Delete root directory & full load.")
     } else {
-      println("Delete root directory & full load.")
-      mssparkutils.fs.rm(adlsDir.concat("/DimensionalModel"), true)
+      if (fileExists(adlsDir.concat("/DimensionalModel"))) {
+        println("Delete root directory & full load.")
+        mssparkutils.fs.rm(adlsDir.concat("/DimensionalModel"), true)
+      } else {
+        println("No root directory.")
+      }
       mssparkutils.fs.put(filePath, accountId)
     }
   }
